refactor: extract CLI commands into focused modules
Major refactoring to break down large CLI files into manageable components:

CLI Commands Refactoring:
- Extract ConfigCommands (~250 lines) into cli_modules/commands/config_commands.py
- Extract AuthCommands (~390 lines) into cli_modules/commands/auth_commands.py
- Reduce cli_commands.py from 981 to 343 lines (65% reduction)
- Maintain backward compatibility through delegator methods

File Organization:
- Rename cli/ to cli_modules/ to avoid namespace conflicts
- Group related commands into focused classes following SRP
- Create proper module structure with __init__.py files

Core Extraction (from ensemble_execution.py):
- Extract ModelFactory (~200 lines) into core/models/model_factory.py
- Extract DependencyAnalyzer (~70 lines) into core/execution/dependency_analyzer.py
- Extract AgentExecutor (~270 lines) into core/execution/agent_executor.py
- Extract InputEnhancer (~70 lines) into core/execution/input_enhancer.py
- Reduce ensemble_execution.py from 1,140 to 978 lines

Benefits:
- Single Responsibility Principle: Each class focuses on specific domain
- Better maintainability: Easier to find and modify functionality
- Improved testability: Focused components are easier to test
- Cleaner imports: Reduced coupling between modules

Linting passes, some tests need updates for new structure.

Addresses #45
