+++ src/llm_orc/cli_modules/commands/__init__.py
@@ -1,0 +1,6 @@
+"""CLI command modules."""
+
+from llm_orc.cli_modules.commands.auth_commands import AuthCommands
+from llm_orc.cli_modules.commands.config_commands import ConfigCommands
+
+__all__ = ["AuthCommands", "ConfigCommands"]
